refactor(watermark): move computeWatermarkConfigHash from dtos to core watermark-hash module
